+++ tests/test_db_setup.py
@@ -1,14 +1,9 @@
-import sys
-
 import pytest
 
 from pytest_django.lazy_django import get_django_version
 from pytest_django_test.db_helpers import (db_exists, drop_database,
                                            mark_database, mark_exists,
                                            skip_if_sqlite_in_memory)
-
-skip_on_python32 = pytest.mark.skipif(sys.version_info[:2] == (3, 2),
-                                      reason='xdist is flaky with Python 3.2')
 
 
 def test_db_reuse_simple(django_testdir):
@@ -87,18 +82,13 @@ def test_db_reuse(django_testdir):
 
 class TestSqlite:
 
-    db_name_17 = 'test_db_name_django17'
-    db_name_before_17 = 'test_db_name_before_django17'
-
     db_settings = {'default': {
         'ENGINE': 'django.db.backends.sqlite3',
         'NAME': 'db_name',
+        'TEST': {
+            'NAME': 'test_custom_db_name'
+        }
     }}
-    from django import VERSION
-    if VERSION > (1, 7):
-        db_settings['default']['TEST'] = {'NAME': db_name_17}
-    else:
-        db_settings['default']['TEST_NAME'] = db_name_before_17
 
     def test_sqlite_test_name_used(self, django_testdir):
 
@@ -113,18 +103,14 @@ class TestSqlite:
 
                 assert conn.vendor == 'sqlite'
                 print(conn.settings_dict)
-                if VERSION > (1,7):
-                    assert conn.settings_dict['NAME'] == '%s'
-                else:
-                    assert conn.settings_dict['NAME'] == '%s'
-        ''' % (self.db_name_17, self.db_name_before_17))
+                assert conn.settings_dict['NAME'] == 'test_custom_db_name'
+        ''')
 
         result = django_testdir.runpytest_subprocess('--tb=short', '-v')
         assert result.ret == 0
         result.stdout.fnmatch_lines(['*test_a*PASSED*'])
 
 
-@skip_on_python32
 def test_xdist_with_reuse(django_testdir):
     skip_if_sqlite_in_memory()
 
@@ -192,8 +178,6 @@ def test_xdist_with_reuse(django_testdir):
 
 class TestSqliteWithXdist:
 
-    pytestmark = skip_on_python32
-
     db_settings = {'default': {
         'ENGINE': 'django.db.backends.sqlite3',
         'NAME': '/tmp/should-not-be-used',
@@ -229,224 +213,14 @@ def test_initial_data(django_testdir_initial):
         from .app.models import Item
 
         @pytest.mark.django_db
-        def test_inner_south():
+        def test_inner():
             assert [x.name for x in Item.objects.all()] \
                 == ["mark_initial_data"]
     ''')
 
     result = django_testdir_initial.runpytest_subprocess('--tb=short', '-v')
     assert result.ret == 0
-    result.stdout.fnmatch_lines(['*test_inner_south*PASSED*'])
-
-
-# NOTE: South tries to monkey-patch management._commands, which has been
-#       replaced by lru_cache and would cause an AttributeError.
-@pytest.mark.skipif(get_django_version() >= (1, 7),
-                    reason='South does not support Django 1.7+')
-@pytest.mark.skipif(sys.version_info[0] == 3,
-                    reason='South is not properly supported on Python 3')
-class TestSouth:
-    """Test interaction with South, with and without initial_data."""
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_initial_data_south_no_migrations(self, django_testdir_initial):
-        django_testdir_initial.create_test_module('''
-            import pytest
-
-            from .app.models import Item
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                assert [x.name for x in Item.objects.all()] \
-                    == ["mark_initial_data"]
-        ''')
-
-        result = django_testdir_initial.runpytest_subprocess('--tb=short', '-v', '-s')
-        result.stdout.fnmatch_lines_random([
-            "tpkg/test_the_test.py::test_inner_south*",
-            "*PASSED*",
-            "*Destroying test database for alias 'default'...*"])
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_initial_data_south_with_migrations(self, django_testdir_initial):
-        """
-        If migrations exists, there should be an error if they do not create
-        the DB table.
-        """
-        django_testdir_initial.create_test_module('''
-            import pytest
-
-            from .app.models import Item
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                assert [x.name for x in Item.objects.all()] \
-                    == ["mark_initial_data"]
-        ''')
-        django_testdir_initial.mkpydir('tpkg/app/south_migrations')
-
-        result = django_testdir_initial.runpytest_subprocess('--tb=short', '-v', '-s')
-        assert result.ret != 0
-        # Can be OperationalError or DatabaseError (Django 1.4) with sqlite,
-        # ProgrammingError with PostgreSQL.
-        result.stdout.fnmatch_lines([
-            '*Error:*app_item*'])
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_initial_south_migrations(self, django_testdir_initial):
-        """This should fail, because it has no real migration that
-        would create the table, and so no initial data can be loaded."""
-        testdir = django_testdir_initial
-        testdir.create_test_module('''
-            import pytest
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                pass
-            ''')
-
-        testdir.create_initial_south_migration()
-
-        result = testdir.runpytest_subprocess('--tb=short', '-v', '-s')
-        assert result.ret == 0
-        result.stdout.fnmatch_lines(['*mark_south_migration_forwards*'])
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_initial_without_south_migrations(self, django_testdir_initial):
-        """Using South, but without any migrations should still load the
-        initial data."""
-        django_testdir_initial.create_test_module('''
-            import pytest
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                pass
-            ''')
-
-        result = django_testdir_initial.runpytest_subprocess('--tb=short', '-v', '-s')
-        assert result.ret == 0
-        result.stdout.fnmatch_lines(['*PASSED*'])
-        assert 'mark_south_migration_forwards' not in result.stdout.str()
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_south_migrations(self, django_testdir):
-        """South migration with a normal testdir (no initial data)."""
-        testdir = django_testdir
-        testdir.create_test_module('''
-            import pytest
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                pass
-            ''')
-
-        testdir.mkpydir('tpkg/app/south_migrations')
-        testdir.create_app_file("""
-            from south.v2 import SchemaMigration
-
-            class Migration(SchemaMigration):
-                def forwards(self, orm):
-                    print("mark_south_migration_forwards")
-            """, 'south_migrations/0001_initial.py')
-        result = testdir.runpytest_subprocess('--tb=short', '-v', '-s')
-        assert result.ret == 0
-        result.stdout.fnmatch_lines(['*mark_south_migration_forwards*'])
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = False
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_south_no_migrations(self, django_testdir_initial):
-        testdir = django_testdir_initial
-        testdir.create_test_module('''
-            import pytest
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                pass
-        ''')
-
-        testdir.mkpydir('tpkg/app/south_migrations')
-        p = testdir.tmpdir.join(
-            "tpkg/app/south_migrations/0001_initial").new(ext="py")
-        p.write('raise Exception("This should not get imported.")',
-                ensure=True)
-
-        result = testdir.runpytest_subprocess('--tb=short', '-v', '-s')
-        assert result.ret == 0
-        result.stdout.fnmatch_lines_random([
-            "tpkg/test_the_test.py::test_inner_south*",
-            "*PASSED*",
-            "*Destroying test database for alias 'default'...*"])
-
-    @pytest.mark.django_project(extra_settings="""
-        INSTALLED_APPS += [ 'south', ]
-        SOUTH_TESTS_MIGRATE = True
-        SOUTH_MIGRATION_MODULES = {
-            'app': 'tpkg.app.south_migrations',
-        }
-        """)
-    def test_south_migrations_python_files_star(self, django_testdir_initial):
-        """
-        Test for South migrations and tests imported via `*.py`.
-
-        This is meant to reproduce
-        https://github.com/pytest-dev/pytest-django/issues/158, but does not
-        fail.
-        """
-        testdir = django_testdir_initial
-        testdir.create_test_module('''
-            import pytest
-
-            @pytest.mark.django_db
-            def test_inner_south():
-                pass
-        ''', 'test.py')
-        testdir.create_initial_south_migration()
-
-        pytest_ini = testdir.create_test_module("""
-            [pytest]
-            python_files=*.py""", 'pytest.ini')
-
-        result = testdir.runpytest_subprocess('--tb=short', '-v', '-s', '-c', pytest_ini)
-        assert result.ret == 0
-        result.stdout.fnmatch_lines_random([
-            "tpkg/test.py::test_inner_south*",
-            "*mark_south_migration_forwards*",
-            "*PASSED*"])
+    result.stdout.fnmatch_lines(['*test_inner*PASSED*'])
 
 
 class TestNativeMigrations(object):